% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_load_infoworks_output.R,
%   R/infoworks_gerris_functions.R
\name{load_infoworks_output}
\alias{load_infoworks_output}
\title{Load the Infoworks Output file (csv)}
\usage{
load_infoworks_output(iw_output_folder, simulation_name)

load_infoworks_output(iw_output_folder, simulation_name)
}
\arguments{
\item{iw_output_folder}{Folder with InfoWorks csv-outpu}

\item{simulation_name}{}
}
\value{
A list with all read csv Files


}
\description{
The function looks up all files in the specified infoWorks output
folder. It chooses csv files that start with "link" and contain "ds" for
donwstream and the specified simulation name. All parameter listed in the
parameter conversion table are selected.
}
